pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "serology-imputation"
version = "0.1.0"
description = "End-to-end serology missing-value imputation (RF IterativeImputer + PyTorch AE) with parallel execution."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Miad Boodaghidizaji" }]
dependencies = [
  "numpy",
  "pandas",
  "pyyaml",
  "scikit-learn",
  "torch",
  "ray",
  "rich",
  "tqdm",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
  "mypy",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
serology-impute = "serology_imputation.run:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"
